Move exceptions into its own module
We already have quite a few exceptions, so we better move them to keep
the base classes cleaner.
