+++ src/frequenz/channels/_base_classes.py
@@ -6,103 +6,12 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, Callable, Generic, Optional, TypeVar
+from typing import Callable, Generic, Optional, TypeVar
+
+from ._exceptions import ReceiverStoppedError
 
 T = TypeVar("T")
 U = TypeVar("U")
-
-
-class Error(RuntimeError):
-    """Base error.
-
-    All exceptions generated by this library inherit from this exception.
-    """
-
-    def __init__(self, message: Any):
-        """Create a ChannelError instance.
-
-        Args:
-            message: An error message.
-        """
-        super().__init__(message)
-
-
-class ChannelError(Error):
-    """An error produced in a channel.
-
-    All exceptions generated by channels inherit from this exception.
-    """
-
-    def __init__(self, message: Any, channel: Any):
-        """Create a ChannelError instance.
-
-        Args:
-            message: An error message.
-            channel: A reference to the channel that encountered the error.
-        """
-        super().__init__(message)
-        self.channel: Any = channel
-
-
-class ChannelClosedError(ChannelError):
-    """Error raised when trying to operate on a closed channel."""
-
-    def __init__(self, channel: Any):
-        """Create a `ChannelClosedError` instance.
-
-        Args:
-            channel: A reference to the channel that was closed.
-        """
-        super().__init__(f"Channel {channel} was closed", channel)
-
-
-class SenderError(Error, Generic[T]):
-    """An error produced in a [Sender][frequenz.channels.Sender].
-
-    All exceptions generated by senders inherit from this exception.
-    """
-
-    def __init__(self, message: Any, sender: Sender[T]):
-        """Create an instance.
-
-        Args:
-            message: An error message.
-            sender: The [Sender][frequenz.channels.Sender] where the error
-                happened.
-        """
-        super().__init__(message)
-        self.sender: Sender[T] = sender
-
-
-class ReceiverError(Error, Generic[T]):
-    """An error produced in a [Receiver][frequenz.channels.Receiver].
-
-    All exceptions generated by receivers inherit from this exception.
-    """
-
-    def __init__(self, message: Any, receiver: Receiver[T]):
-        """Create an instance.
-
-        Args:
-            message: An error message.
-            receiver: The [Receiver][frequenz.channels.Receiver] where the
-                error happened.
-        """
-        super().__init__(message)
-        self.receiver: Receiver[T] = receiver
-
-
-class ReceiverStoppedError(ReceiverError[T]):
-    """The [Receiver][frequenz.channels.Receiver] stopped producing messages."""
-
-    def __init__(self, receiver: Receiver[T]):
-        """Create an instance.
-
-        Args:
-            receiver: The [Receiver][frequenz.channels.Receiver] where the
-                error happened.
-        """
-        super().__init__(f"Receiver {receiver} was stopped", receiver)
 
 
 class Sender(ABC, Generic[T]):
